% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lines.R
\name{create_lines}
\alias{create_lines}
\title{Create script lines}
\usage{
create_lines(from, to, action, color, sleep_time_secs)
}
\description{
A cursor always go:
}
\details{
\itemize{
\item movement, \link{create_movement_lines}
\item select
\item movement, \link{create_movement_lines}
\item do action
\item movement, \link{create_movement_lines}
}

The cursor will always end up back at the original position
}
