% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_movement_lines.R
\name{create_movement_lines}
\alias{create_movement_lines}
\title{Create the movement lines in a script}
\usage{
create_movement_lines(from, to, color)
}
\description{
Create the movement lines in a script
}
