% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_script_text.R
\name{create_script_text}
\alias{create_script_text}
\title{Create the text of a script}
\usage{
create_script_text(sleep_time_secs = 8)
}
\description{
Create the text of a script
}
\note{
Cursors always move back to their original position

\if{html}{\out{<div class="sourceCode">}}\preformatted{[Event "It (cat.17)"]
[Site "Wijk aan Zee (Netherlands)"]
[Date "1999.??.??"]
[Round "?"]
[White "Garry Kasparov"]
[Black "Veselin Topalov"]
[Result "1-0"]
[TimeControl ""]
[Link "https://www.chess.com/games/view/969971"]
}\if{html}{\out{</div>}}
}
