% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_vertical_lines.R
\name{create_vertical_lines}
\alias{create_vertical_lines}
\title{Create script lines}
\usage{
create_vertical_lines(from = "e2", to = "e4", color = "white")
}
\description{
A cursor always go:
}
\details{
\itemize{
\item horizontal, \link{create_horizontal_lines}
\item vertical
\item select
\item horizontal
\item vertical
\item do action
\item horizontal
\item vertical
}

The cursor will always end up back at the original position
}
